+++ models/experimental/functional_unet/tests/test_unet_model.py
@@ -4,14 +4,15 @@
 
 import pytest
 import ttnn
-from loguru import logger
 
 from tests.ttnn.utils_for_testing import assert_with_pcc
 
-from models.experimental.functional_unet.unet_utils import create_unet_input_tensors
+from models.experimental.functional_unet.tt.model_preprocessing import (
+    create_unet_input_tensors,
+    create_unet_model_parameters,
+)
 from models.experimental.functional_unet.tt import unet_shallow_torch
-from models.experimental.functional_unet.tt import unet_shallow_ttnn2
-from models.experimental.functional_unet.tt import model_preprocessing
+from models.experimental.functional_unet.tt import unet_shallow_ttnn
 
 
 @pytest.mark.parametrize("batch", [2])
@@ -21,8 +22,8 @@ def test_unet_model(batch, groups, device, use_program_cache):
     torch_input, ttnn_input = create_unet_input_tensors(device, batch, groups, pad_input=True)
     model = unet_shallow_torch.UNet.from_random_weights(groups=1)
 
-    parameters = model_preprocessing.create_unet_model_parameters(model, torch_input, groups=groups, device=device)
-    ttnn_model = unet_shallow_ttnn2.UNet(parameters, device)
+    parameters = create_unet_model_parameters(model, torch_input, groups=groups, device=device)
+    ttnn_model = unet_shallow_ttnn.UNet(parameters, device)
 
     torch_output_tensor = model(torch_input)
     output_tensor = ttnn_model(ttnn_input, list(torch_input.shape))
