+++ models/experimental/functional_unet/tests/test_unet_upblock.py
@@ -2,17 +2,17 @@
 
 # SPDX-License-Identifier: Apache-2.0
 
-import torch
 import pytest
 import ttnn
-from loguru import logger
 
 from tests.ttnn.utils_for_testing import assert_with_pcc
 
-from models.experimental.functional_unet.unet_utils import create_unet_input_tensors
+from models.experimental.functional_unet.tt.model_preprocessing import (
+    create_unet_input_tensors,
+    create_unet_model_parameters,
+)
 from models.experimental.functional_unet.tt import unet_shallow_torch
-from models.experimental.functional_unet.tt import unet_shallow_ttnn2
-from models.experimental.functional_unet.tt import model_preprocessing
+from models.experimental.functional_unet.tt import unet_shallow_ttnn
 
 
 def check_pcc_conv(torch_tensor, ttnn_tensor, pcc=0.995):
@@ -36,8 +36,8 @@ def test_unet_upblock(batch, groups, block_name, input_channels, input_height, i
     torch_input, ttnn_input = create_unet_input_tensors(device, batch, groups, pad_input=False)
     model = unet_shallow_torch.UNet.from_random_weights(groups=groups)
 
-    parameters = model_preprocessing.create_unet_model_parameters(model, torch_input, groups=groups, device=device)
-    ttnn_model = unet_shallow_ttnn2.UNet(parameters, device)
+    parameters = create_unet_model_parameters(model, torch_input, groups=groups, device=device)
+    ttnn_model = unet_shallow_ttnn.UNet(parameters, device)
 
     torch_input, ttnn_input = create_unet_input_tensors(
         device,
