+++ models/experimental/functional_unet/tt/model_preprocessing.py
@@ -3,10 +3,36 @@
 # SPDX-License-Identifier: Apache-2.0
 
 import torch
+import ttnn
+
 from ttnn.model_preprocessing import infer_ttnn_module_args
 
 from models.experimental.functional_unet.tt import unet_shallow_torch
-from models.experimental.functional_unet.tt import unet_shallow_ttnn2
+
+
+def create_unet_input_tensors(
+    device, batch, groups, pad_input=True, input_channels=4, input_height=1056, input_width=160
+):
+    torch_input_tensor = torch.randn(batch, input_channels * groups, input_height, input_width)
+    ttnn_input_tensor = torch.permute(torch_input_tensor, (0, 2, 3, 1))
+    ttnn_input_tensor = ttnn_input_tensor.reshape(
+        1,
+        1,
+        ttnn_input_tensor.shape[0] * ttnn_input_tensor.shape[1] * ttnn_input_tensor.shape[2],
+        ttnn_input_tensor.shape[3],
+    )
+    if pad_input:
+        # Pad to 16 if grayskull run and 32 for wormhole
+        pad = 32 if device.arch() == ttnn.device.Arch.WORMHOLE_B0 else 16
+        hpad = 0  # 96*32*64
+        if ttnn_input_tensor.shape[-1] < pad or ttnn_input_tensor.shape[-2] < hpad:
+            ttnn_input_tensor = torch.nn.functional.pad(
+                ttnn_input_tensor,
+                (0, max(0, pad - ttnn_input_tensor.shape[-1]), 0, max(0, hpad - ttnn_input_tensor.shape[-2])),
+            )
+    ttnn_input_tensor = ttnn.from_torch(ttnn_input_tensor, dtype=ttnn.bfloat16)
+
+    return torch_input_tensor, ttnn_input_tensor
 
 
 def create_unet_model_parameters(model: unet_shallow_torch.UNet, input_tensor: torch.Tensor, groups: int, device):
