+++ models/experimental/functional_unet/tt/unet_shallow_ttnn.py
@@ -3,9 +3,14 @@
 # SPDX-License-Identifier: Apache-2.0
 
 import ttnn
+import torch
 
-from tt_lib import profiler
 from loguru import logger
+
+from ttnn.operations.conv2d import determine_parallel_config, create_sharded_memory_config_from_parallel_config
+
+from models.utility_functions import nearest_32
+from ttnn.model_preprocessing import fold_batch_norm2d_into_conv2d, ParameterDict
 
 
 # Unet reshard wrapper
@@ -62,7 +67,7 @@ def unet_concat(ttnn_tensors, dim=-1, use_reshard=True, perf_mode=False):
                 t_mem_config.shard_spec.shape = reshard_shape
                 t_mem_config.shard_spec.grid = output_mem_config.shard_spec.grid
                 t_mem_config.shard_spec.orientation = output_mem_config.shard_spec.orientation
-                ttlib_tensors[i] = ttnn.reshard(t, t_mem_config)
+                ttlib_tensors[i] = ttnn.experimental.tensor.reshard(t, t_mem_config)
     else:
         output_mem_config = ttnn.DRAM_MEMORY_CONFIG
         for i in range(0, len(ttlib_tensors)):
@@ -71,199 +76,421 @@ def unet_concat(ttnn_tensors, dim=-1, use_reshard=True, perf_mode=False):
     return ttnn.concat(ttlib_tensors, dim=dim, memory_config=output_mem_config)
 
 
-# Unet shallow ttnn implementation
-class UNet:
+class UNetConv2D:
     def __init__(
         self,
-        parameters,
-    ) -> None:
-        self.c1 = parameters.c1
-        self.c1_2 = parameters.c1_2
-        self.p1 = parameters.p1
-        self.c2 = parameters.c2
-        self.c2_2 = parameters.c2_2
-        self.p2 = parameters.p2
-        self.c3 = parameters.c3
-        self.c3_2 = parameters.c3_2
-        self.p3 = parameters.p3
-        self.c4 = parameters.c4
-        self.c4_2 = parameters.c4_2
-        self.p4 = parameters.p4
-        self.bnc = parameters.bnc
-        self.bnc_2 = parameters.bnc_2
-        self.c5 = parameters.c5
-        self.c5_2 = parameters.c5_2
-        self.c5_3 = parameters.c5_3
-        self.c6 = parameters.c6
-        self.c6_2 = parameters.c6_2
-        self.c6_3 = parameters.c6_3
-        self.c7 = parameters.c7
-        self.c7_2 = parameters.c7_2
-        self.c7_3 = parameters.c7_3
-        self.c8 = parameters.c8
-        self.c8_2 = parameters.c8_2
-        self.c8_3 = parameters.c8_3
-        self.output_layer = parameters.output_layer
-        self.cache = {}
+        conv,
+        bn=None,
+        device=None,
+        cache={},
+        activation="relu",
+        activation_dtype=ttnn.bfloat8_b,
+        weights_dtype=ttnn.bfloat8_b,
+    ):
+        self.device = device
+        self.batch_size = conv.batch_size
+        self.input_height = conv.input_height
+        self.input_width = conv.input_width
+        self.in_channels = conv.in_channels
+        self.out_channels = conv.out_channels
+        self.kernel_size = conv.kernel_size
+        self.padding = conv.padding
+        self.stride = conv.stride
+        self.groups = conv.groups
+        self.use_1d_systolic_array = conv.use_1d_systolic_array
+        self.deallocate_activation = True
+        self.cache = cache
 
-    def __call__(self, device, input_tensor, original_shape, perf_mode=False):
-        nhw = original_shape[-4] * original_shape[-2] * original_shape[-1]
-        input_tensor = input_tensor.to(device, self.c1.conv.input_sharded_memory_config)
+        self.conv_config = ttnn.Conv2dConfig(
+            dtype=activation_dtype,
+            weights_dtype=weights_dtype,
+            math_fidelity=ttnn.MathFidelity.LoFi,
+            height_sharding=self.use_1d_systolic_array,
+            deallocate_activation=self.deallocate_activation,
+            fp32_dest_acc_enabled=True,
+            packer_l1_accum_enabled=False,
+            enable_act_double_buffer=False,
+            enable_split_reader=False,
+            enable_subblock_padding=False,
+            activation=activation,
+        )
+        config_override = conv.conv_blocking_and_parallelization_config_override
+        if config_override and "act_block_h" in config_override:
+            self.conv_config.act_block_h_override = config_override["act_block_h"]
 
-        profiler.tracy_message("c1")
-        output_tensor = self.c1(input_tensor)
-        output_tensor = self.c1_2(output_tensor)
-        if perf_mode:
-            save_c1_2_out = ttnn.to_layout(output_tensor, layout=ttnn.ROW_MAJOR_LAYOUT)
+        if bn is not None:
+            weight, bias = fold_batch_norm2d_into_conv2d(conv.module, bn.module)
         else:
-            save_c1_2_out = ttnn.sharded_to_interleaved(
-                output_tensor, ttnn.DRAM_MEMORY_CONFIG, output_dtype=ttnn.experimental.tensor.DataType.BFLOAT16
-            )
-        output_tensor = self.p1(output_tensor)
+            weight, bias = conv.module.weight, conv.module.bias
 
-        profiler.tracy_message("c2")
-        output_tensor = unet_reshard(output_tensor, self.c2.conv.input_sharded_memory_config, use_reshard=False)
-        output_tensor = self.c2(output_tensor)
-        output_tensor = self.c2_2(output_tensor)
-        if perf_mode:
-            save_c2_2_out = output_tensor
-        else:
-            save_c2_2_out = ttnn.sharded_to_interleaved(
-                output_tensor, ttnn.DRAM_MEMORY_CONFIG, output_dtype=ttnn.experimental.tensor.DataType.BFLOAT16
-            )
-        output_tensor = self.p2(output_tensor)
+        weight = weight
+        bias = torch.reshape(bias, (1, 1, 1, -1))
 
-        profiler.tracy_message("c3")
-        output_tensor = unet_reshard(output_tensor, self.c3.conv.input_sharded_memory_config)
-        output_tensor = self.c3(output_tensor)
-        output_tensor = self.c3_2(output_tensor)
-        if perf_mode:
-            save_c3_2_out = output_tensor
-        else:
-            save_c3_2_out = ttnn.sharded_to_interleaved(
-                output_tensor, ttnn.DRAM_MEMORY_CONFIG, output_dtype=ttnn.experimental.tensor.DataType.BFLOAT16
-            )
-        output_tensor = self.p3(output_tensor)
+        if bias.shape[-1] == 1:
+            bias = bias.repeat((1, 1, 32, 32))
+            logger.warning(f"Found a bias we need to replicate {bias.shape}")
 
-        profiler.tracy_message("c4")
-        output_tensor = unet_reshard(output_tensor, self.c4.conv.input_sharded_memory_config)
-        output_tensor = self.c4(output_tensor)
-        output_tensor = self.c4_2(output_tensor)
-        if perf_mode:
-            save_c4_2_out = output_tensor
-        else:
-            save_c4_2_out = ttnn.sharded_to_interleaved(
-                output_tensor, ttnn.DRAM_MEMORY_CONFIG, output_dtype=ttnn.experimental.tensor.DataType.BFLOAT16
-            )
-        output_tensor = self.p4(output_tensor)
+        self.weight = ttnn.from_torch(weight, dtype=ttnn.float32)
+        self.bias = ttnn.from_torch(bias, dtype=ttnn.float32)
 
-        profiler.tracy_message("bnc")
-        output_tensor = unet_reshard(output_tensor, self.bnc.conv.input_sharded_memory_config)
-        output_tensor = self.bnc(output_tensor)
-        output_tensor = self.bnc_2(output_tensor)
-
-        if not perf_mode:
-            # need to convert into interleaved, then back into sharded due to pcc issues
-            # for certain tensor shape sizes, you get pcc issues when trying to convert between data layouts
-            output_tensor = ttnn.sharded_to_interleaved(output_tensor, ttnn.DRAM_MEMORY_CONFIG)
-        output_tensor = ttnn.to_layout(output_tensor, layout=ttnn.ROW_MAJOR_LAYOUT)
-
-        ## upsample block
-        profiler.tracy_message("upsample1")
-        output_tensor = ttnn.to_memory_config(output_tensor, memory_config=ttnn.DRAM_MEMORY_CONFIG)
-        output_tensor = ttnn.upsample(output_tensor, (2, 2, 1))
-        output_tensor = ttnn.reshape(output_tensor, (1, 1, nhw // 64, output_tensor.shape[-1]))
-
-        profiler.tracy_message("concat1")
-        save_c4_2_out = ttnn.to_layout(save_c4_2_out, layout=ttnn.ROW_MAJOR_LAYOUT)
-        output_tensor = unet_concat([output_tensor, save_c4_2_out], dim=-1, perf_mode=perf_mode)
-
-        profiler.tracy_message("c5")
-        output_tensor = unet_reshard(output_tensor, self.c5.conv.input_sharded_memory_config)
-        output_tensor = self.c5(output_tensor)
-        output_tensor = self.c5_2(output_tensor)
-        output_tensor = self.c5_3(output_tensor)
-
-        if not perf_mode:
-            # need to convert into interleaved, then back into sharded due to pcc issues
-            # for certain tensor shape sizes, you get pcc issues when trying to convert between data layouts
-            output_tensor = ttnn.sharded_to_interleaved(output_tensor, ttnn.DRAM_MEMORY_CONFIG)
-        output_tensor = ttnn.to_layout(output_tensor, layout=ttnn.ROW_MAJOR_LAYOUT)
-
-        profiler.tracy_message("upsample2")
-        output_tensor = ttnn.to_memory_config(output_tensor, memory_config=ttnn.DRAM_MEMORY_CONFIG)
-        output_tensor = ttnn.upsample(output_tensor, (2, 2, 1))
-        output_tensor = ttnn.reshape(output_tensor, (1, 1, nhw // 16, output_tensor.shape[-1]))
-
-        profiler.tracy_message("concat2")
-        save_c3_2_out = ttnn.to_layout(save_c3_2_out, layout=ttnn.ROW_MAJOR_LAYOUT)
-        output_tensor = unet_concat([output_tensor, save_c3_2_out], dim=-1, perf_mode=perf_mode)
-
-        profiler.tracy_message("c6")
-        output_tensor = unet_reshard(output_tensor, self.c6.conv.input_sharded_memory_config)
-        output_tensor = self.c6(output_tensor)
-        output_tensor = self.c6_2(output_tensor)
-        output_tensor = self.c6_3(output_tensor)
-
-        if not perf_mode:
-            # need to convert into interleaved, then back into sharded due to pcc issues
-            # for certain tensor shape sizes, you get pcc issues when trying to convert between data layouts
-            output_tensor = ttnn.sharded_to_interleaved(output_tensor, ttnn.DRAM_MEMORY_CONFIG)
-        output_tensor = ttnn.to_layout(output_tensor, layout=ttnn.ROW_MAJOR_LAYOUT)
-
-        profiler.tracy_message("upsample3")
-        output_tensor = ttnn.to_memory_config(output_tensor, memory_config=ttnn.DRAM_MEMORY_CONFIG)
-        output_tensor = ttnn.upsample(output_tensor, (2, 2, 1))
-        output_tensor = ttnn.reshape(output_tensor, (1, 1, nhw // 4, output_tensor.shape[-1]))
-
-        profiler.tracy_message("concat3")
-        save_c2_2_out = ttnn.to_layout(save_c2_2_out, layout=ttnn.ROW_MAJOR_LAYOUT)
-        output_tensor = unet_concat([output_tensor, save_c2_2_out], dim=-1, perf_mode=perf_mode)
-
-        profiler.tracy_message("c7")
-        hacked_shard_shape = self.c7.conv.input_sharded_memory_config.shard_spec.shape
-        hacked_shard_shape[1] = output_tensor.shape[-1]
-        self.c7.conv.input_sharded_memory_config.shard_spec.shape = hacked_shard_shape
-        output_tensor = unet_reshard(output_tensor, self.c7.conv.input_sharded_memory_config)
-        output_tensor = self.c7(output_tensor)
-        output_tensor = self.c7_2(output_tensor)
-        output_tensor = self.c7_3(output_tensor)
-
-        if not perf_mode:
-            # need to convert into interleaved, then back into sharded due to pcc issues
-            # for certain tensor shape sizes, you get pcc issues when trying to convert between data layouts
-            output_tensor = ttnn.sharded_to_interleaved(output_tensor, ttnn.DRAM_MEMORY_CONFIG)
-        output_tensor = ttnn.to_layout(output_tensor, layout=ttnn.ROW_MAJOR_LAYOUT)
-
-        profiler.tracy_message("upsample4")
-        output_tensor = ttnn.to_memory_config(output_tensor, memory_config=ttnn.DRAM_MEMORY_CONFIG)
-        output_tensor = ttnn.upsample(output_tensor, (2, 2, 1))
-        output_tensor = ttnn.reshape(output_tensor, (1, 1, nhw, output_tensor.shape[-1]))
-
-        profiler.tracy_message("concat4")
-        output_tensor = unet_concat([output_tensor, save_c1_2_out], dim=-1, perf_mode=perf_mode)
-
-        profiler.tracy_message("c8")
-        output_tensor = unet_reshard(output_tensor, self.c8.conv.input_sharded_memory_config)
-        output_tensor = self.c8(output_tensor)
-        output_tensor = self.c8_2(output_tensor)
-        output_tensor = self.c8_3(output_tensor)
-
-        output_tensor = output_tensor.cpu().pad_to_tile(0)
-        output_tensor, _, _, self.output_layer.conv.weight, self.output_layer.conv.bias = ttnn.conv2d(
-            input_tensor=output_tensor,
-            weight_tensor=self.output_layer.conv.weight,
-            bias_tensor=self.output_layer.conv.bias,
-            in_channels=self.output_layer.in_channels,
-            out_channels=self.output_layer.out_channels,
-            input_height=self.output_layer.input_height,
-            input_width=self.output_layer.input_width,
-            batch_size=self.output_layer.batch_size,
-            kernel_size=(1, 1),
-            stride=(1, 1),
-            padding=(0, 0),
+    def __call__(self, x):
+        logger.info(
+            f"running conv - input={x.shape} in_h={self.input_height}, in_w={self.input_width}, in_c={self.in_channels}, out=c{self.out_channels}, pad={self.padding}, weight={self.weight.shape}"
+        )
+        x, output_height, output_width, self.weight, self.bias = ttnn.conv2d(
+            input_tensor=x,
+            weight_tensor=self.weight,
+            bias_tensor=self.bias,
+            device=self.device,
+            in_channels=self.in_channels,
+            out_channels=self.out_channels,
+            input_height=self.input_height,
+            input_width=self.input_width,
+            batch_size=self.batch_size,
+            kernel_size=self.kernel_size,
+            stride=self.stride,
+            padding=self.padding,
+            conv_config=self.conv_config,
             conv_op_cache=self.cache,
+        )
+        logger.info(f"done conv out shape: {output_height} {output_width}")
+        return x
+
+
+class UNetMaxPool2D:
+    def __init__(self, pool, device=None, reader_patterns_cache={}):
+        self.pool = pool
+        self.max_pool = ttnn.MaxPool2d(
+            kernel_size=pool.kernel_size,
+            stride=pool.stride,
+            padding=pool.padding,
+            dilation=pool.dilation,
+            dtype=pool.dtype,
+            batch_size=pool.batch_size,
+            input_height=pool.input_height,
+            input_width=pool.input_width,
+            reader_patterns_cache=reader_patterns_cache,
+            parallel_config_override=pool.parallel_config_override,
+            deallocate_activation=True,
             device=device,
         )
 
-        return ttnn.from_device(output_tensor)
+    def __call__(self, x):
+        # For some reason the shard widths don't always match - so don't assert on it
+        assert (
+            x.memory_config().shard_spec.num_cores()
+            == self.max_pool.max_pool.input_sharded_memory_config.shard_spec.num_cores()
+            and x.memory_config().shard_spec.shape[0]
+            == self.max_pool.max_pool.input_sharded_memory_config.shard_spec.shape[0]
+        ), "Expected same input shard to match max pool's shard configuration"
+        return self.max_pool(x)
+
+
+class UNetMaxPool2DNew:
+    def __init__(self, pool, channels, device=None, reader_patterns_cache={}):
+        self.channels = channels
+        self.kernel_size = pool.kernel_size
+        self.stride = pool.stride
+        self.padding = pool.padding
+        self.dilation = pool.dilation
+        self.batch_size = pool.batch_size
+        self.input_height = pool.input_height
+        self.input_width = pool.input_width
+        self.device = device
+
+    def __call__(self, x):
+        return ttnn.max_pool2d_new(
+            input_tensor=x,
+            batch_size=self.batch_size,
+            input_h=self.input_height,
+            input_w=self.input_width,
+            channels=self.channels,
+            kernel_size=[self.kernel_size, self.kernel_size],
+            stride=[self.stride, self.stride],
+            padding=[self.padding, self.padding],
+            dilation=[self.dilation, self.dilation],
+            device=self.device,
+        )
+
+
+class UNetDownblock:
+    def __init__(self, conv1, bn1, conv2, bn2, pool, device, conv_cache={}, max_pool_cache={}, should_reshard=False):
+        self.conv1 = UNetConv2D(conv1, bn=bn1, device=device, cache=conv_cache)
+        self.conv2 = UNetConv2D(conv2, bn=bn2, device=device, cache=conv_cache)
+        self.pool1 = UNetMaxPool2D(pool, device=device, reader_patterns_cache=max_pool_cache)
+
+        self.should_reshard = should_reshard
+        if self.should_reshard:
+            self.parallel_config = determine_parallel_config(
+                is_1d_systolic=True,
+                batch_size=self.conv1.batch_size,
+                input_channels=self.conv1.in_channels,
+                output_height=self.conv2.input_height,
+                output_width=self.conv2.input_width,
+                output_channels=self.conv1.out_channels,
+                device=device,
+                is_out_tiled=True,
+            )
+
+    def __call__(self, x):
+        assert list(x.shape) == [
+            1,
+            1,
+            nearest_32(self.conv1.input_height * self.conv1.input_width * self.conv1.batch_size),
+            x.shape[-1],  # Channels can be padded
+        ], f"Expected downblock input to flattened into [1, 1, BHW, C] but was {list(x.shape)}"
+        if self.should_reshard:
+            sharded_memory_config = create_sharded_memory_config_from_parallel_config(
+                tensor_shape=[
+                    1,
+                    1,
+                    x.shape[0] * x.shape[1] * x.shape[2],
+                    x.shape[3],
+                ],
+                parallel_config=self.parallel_config,
+                tile_size=32 if x.dtype == ttnn.bfloat8_b else 1,
+            )
+            x = ttnn.to_memory_config(
+                x,
+                memory_config=sharded_memory_config,
+            )
+        x = self.conv1(x)
+        x = self.conv2(x)
+        residual = ttnn.sharded_to_interleaved(
+            x, memory_config=ttnn.DRAM_MEMORY_CONFIG, output_dtype=ttnn.bfloat16
+        )  # pool deletes its activation - spill to DRAM
+        x = self.pool1(x)
+        return x, residual
+
+
+class UNetUpblock:
+    def __init__(self, conv1, bn1, conv2, bn2, conv3, bn3, device, conv_cache={}, should_reshard=False):
+        self.device = device
+        self.conv1 = UNetConv2D(conv1, bn1, device, conv_cache)
+        self.conv2 = UNetConv2D(conv2, bn2, device, conv_cache)
+        self.conv3 = UNetConv2D(conv3, bn3, device, conv_cache)
+
+        self.should_reshard = should_reshard
+        if self.should_reshard:
+            parallel_config = determine_parallel_config(
+                is_1d_systolic=True,
+                batch_size=self.conv1.batch_size,
+                input_channels=self.conv1.in_channels,
+                output_height=self.conv2.input_height,
+                output_width=self.conv2.input_width,
+                output_channels=self.conv1.out_channels,
+                device=device,
+                is_out_tiled=True,
+            )
+            self.sharded_memory_config = create_sharded_memory_config_from_parallel_config(
+                tensor_shape=[
+                    1,
+                    1,
+                    self.conv1.input_width * self.conv1.input_height * self.conv1.batch_size,
+                    nearest_32(self.conv1.in_channels),
+                ],
+                parallel_config=parallel_config,
+                tile_size=32 if conv1.dtype == ttnn.bfloat8_b else 1,
+            )
+            logger.info(f"Created shardspec: {parallel_config}, {self.sharded_memory_config}")
+
+    def upsample(self, x):
+        # Need to reshape into (B, H, W, C) to get correct output from ttnn.upsample
+        x = ttnn.reshape(
+            x, (self.conv1.batch_size, self.conv1.input_height // 2, self.conv1.input_width // 2, x.shape[-1])
+        )
+        x = ttnn.upsample(x, (2, 2, 1))
+        x = ttnn.reshape(
+            x, (1, 1, self.conv1.batch_size * self.conv1.input_height * self.conv1.input_width, x.shape[-1])
+        )
+        return x
+
+    def __call__(self, x, residual):
+        logger.info(
+            f"Running upsample block with {x.shape}, {x.memory_config()}, {residual.shape}, {residual.memory_config()}"
+        )
+        assert list(x.shape)[:2] == [
+            1,
+            1,
+        ], f"Expected downblock input to flattened into [1, 1, BHW, C] but was {list(x.shape)}"
+
+        x = ttnn.to_layout(x, layout=ttnn.ROW_MAJOR_LAYOUT)
+        x = ttnn.to_memory_config(x, memory_config=ttnn.DRAM_MEMORY_CONFIG)
+
+        x = self.upsample(x)
+
+        residual = ttnn.to_layout(residual, ttnn.ROW_MAJOR_LAYOUT)
+        x = unet_concat([x, residual], dim=-1, perf_mode=True)
+        ttnn.deallocate(residual)
+
+        if self.should_reshard:
+            x = ttnn.to_memory_config(x, self.sharded_memory_config)
+
+        x = self.conv1(x)
+        x = self.conv2(x)
+        x = self.conv3(x)
+
+        return x
+
+
+class UNet:
+    def __init__(self, parameters: ParameterDict, device) -> None:
+        self.device = device
+        self.conv_cache = {}
+        self.max_pool_cache = {}
+        self.downblock1 = UNetDownblock(
+            parameters.c1,
+            parameters.b1,
+            parameters.c1_2,
+            parameters.b1_2,
+            parameters.p1,
+            device,
+            conv_cache=self.conv_cache,
+            max_pool_cache=self.max_pool_cache,
+            should_reshard=True,
+        )
+        self.downblock2 = UNetDownblock(
+            parameters.c2,
+            parameters.b2,
+            parameters.c2_2,
+            parameters.b2_2,
+            parameters.p2,
+            device,
+            conv_cache=self.conv_cache,
+            max_pool_cache=self.max_pool_cache,
+            should_reshard=True,
+        )
+        self.downblock3 = UNetDownblock(
+            parameters.c3,
+            parameters.b3,
+            parameters.c3_2,
+            parameters.b3_2,
+            parameters.p3,
+            device,
+            conv_cache=self.conv_cache,
+            max_pool_cache=self.max_pool_cache,
+            should_reshard=True,
+        )
+        self.downblock4 = UNetDownblock(
+            parameters.c4,
+            parameters.b4,
+            parameters.c4_2,
+            parameters.b4_2,
+            parameters.p4,
+            device,
+            conv_cache=self.conv_cache,
+            max_pool_cache=self.max_pool_cache,
+            should_reshard=True,
+        )
+
+        self.bnc = UNetConv2D(parameters.bnc, parameters.bnb, device, cache=self.conv_cache)
+        self.bnc2 = UNetConv2D(parameters.bnc_2, parameters.bnb_2, device, cache=self.conv_cache)
+        bnc_parallel_config = determine_parallel_config(
+            is_1d_systolic=True,
+            batch_size=self.bnc.batch_size,
+            input_channels=self.bnc.in_channels,
+            output_height=self.bnc2.input_height,
+            output_width=self.bnc2.input_width,
+            output_channels=self.bnc.out_channels,
+            device=device,
+            is_out_tiled=True,
+        )
+        self.bnc_sharded_memory_config = create_sharded_memory_config_from_parallel_config(
+            tensor_shape=[
+                1,
+                1,
+                self.bnc.input_width * self.bnc.input_height * self.bnc.batch_size,
+                self.bnc.in_channels,
+            ],
+            parallel_config=bnc_parallel_config,
+            tile_size=(32 if self.bnc.conv_config.dtype == ttnn.bfloat8_b else 1),
+        )
+        logger.info(f"Created bottleneck shard spec: {bnc_parallel_config}, {self.bnc_sharded_memory_config}")
+
+        self.upblock1 = UNetUpblock(
+            parameters.c5,
+            parameters.b5,
+            parameters.c5_2,
+            parameters.b5_2,
+            parameters.c5_3,
+            parameters.b5_3,
+            device,
+            conv_cache=self.conv_cache,
+            should_reshard=False,
+        )
+        self.upblock2 = UNetUpblock(
+            parameters.c6,
+            parameters.b6,
+            parameters.c6_2,
+            parameters.b6_2,
+            parameters.c6_3,
+            parameters.b6_3,
+            device,
+            conv_cache=self.conv_cache,
+            should_reshard=True,
+        )
+        self.upblock3 = UNetUpblock(
+            parameters.c7,
+            parameters.b7,
+            parameters.c7_2,
+            parameters.b7_2,
+            parameters.c7_3,
+            parameters.b7_3,
+            device,
+            conv_cache=self.conv_cache,
+            should_reshard=True,
+        )
+        self.upblock4 = UNetUpblock(
+            parameters.c8,
+            parameters.b8,
+            parameters.c8_2,
+            parameters.b8_2,
+            parameters.c8_3,
+            parameters.b8_3,
+            device,
+            conv_cache=self.conv_cache,
+            should_reshard=True,
+        )
+
+        self.output_layer = UNetConv2D(
+            parameters.output_layer,
+            bn=None,
+            device=device,
+            cache=self.conv_cache,
+            activation="",
+            activation_dtype=ttnn.bfloat16,
+            weights_dtype=ttnn.bfloat8_b,
+        )
+
+    def bottleneck(self, x):
+        x = ttnn.to_memory_config(
+            x,
+            memory_config=self.bnc_sharded_memory_config,
+        )
+        x = self.bnc(x)
+        return self.bnc2(x)
+
+    def __call__(self, x, original_shape, perf_mode=False):
+        assert len(x.shape) == 4, f"Expected UNet input tensors to be rank 4 (was {len(x.shape)})"
+
+        x = x.to(self.device, ttnn.L1_MEMORY_CONFIG)
+
+        x, c1_residual = self.downblock1(x)
+        x, c2_residual = self.downblock2(x)
+        x, c3_residual = self.downblock3(x)
+        x, c4_residual = self.downblock4(x)
+
+        x = self.bottleneck(x)
+
+        x = self.upblock1(x, c4_residual)
+        x = self.upblock2(x, c3_residual)
+        x = self.upblock3(x, c2_residual)
+        x = self.upblock4(x, c1_residual)
+
+        # Pointwise convolutions currently don't handle padded inputs
+        x = x.cpu().pad_to_tile(0)
+        x = self.output_layer(x)
+
+        x = ttnn.from_device(x)
+
+        return x
